+++ cscodec/cscodec.h264/decoder/H264Data.cs
@@ -18,14 +18,14 @@ namespace cscodec.h264.decoder
 			17, 18, 20, 24, 19, 21, 26, 28, 23, 27, 29, 30, 22, 25, 38, 41,
 		};
 
-		public static readonly int[] zigzag_scan = {
+		public static readonly byte[] zigzag_scan = {
 			0+0*4, 1+0*4, 0+1*4, 0+2*4,
 			1+1*4, 2+0*4, 3+0*4, 2+1*4,
 			1+2*4, 0+3*4, 1+3*4, 2+2*4,
 			3+1*4, 3+2*4, 2+3*4, 3+3*4,
 		};
 
-		public static readonly int[] field_scan = {
+		public static readonly byte[] field_scan = {
 			0+0*4, 0+1*4, 1+0*4, 0+2*4,
 			0+3*4, 1+1*4, 1+2*4, 1+3*4,
 			2+0*4, 2+1*4, 2+2*4, 2+3*4,
@@ -46,12 +46,12 @@ namespace cscodec.h264.decoder
 			1*16 + 1*64, 3*16 + 1*64, 1*16 + 3*64, 3*16 + 3*64,
 		};
 
-		public static readonly int[] chroma_dc_scan = {
+		public static readonly byte[] chroma_dc_scan = {
 			0,1,2,3
 		};
 
 		// zigzag_scan8x8_cavlc[i] = zigzag_scan8x8[(i/4) + 16*(i%4)]
-		public static readonly int[] zigzag_scan8x8_cavlc = {
+		public static readonly byte[] zigzag_scan8x8_cavlc = {
 			0+0*8, 1+1*8, 1+2*8, 2+2*8,
 			4+1*8, 0+5*8, 3+3*8, 7+0*8,
 			3+4*8, 1+7*8, 5+3*8, 6+3*8,
@@ -70,7 +70,7 @@ namespace cscodec.h264.decoder
 			5+5*8, 6+5*8, 6+6*8, 7+7*8,
 		};
 
-		public static readonly int[] field_scan8x8 = {
+		public static readonly byte[] field_scan8x8 = {
 			0+0*8, 0+1*8, 0+2*8, 1+0*8,
 			1+1*8, 0+3*8, 0+4*8, 1+2*8,
 			2+0*8, 1+3*8, 0+5*8, 0+6*8,
@@ -89,7 +89,7 @@ namespace cscodec.h264.decoder
 			7+4*8, 7+5*8, 7+6*8, 7+7*8,
 		};
 
-		public static readonly int[] field_scan8x8_cavlc = {
+		public static readonly byte[] field_scan8x8_cavlc = {
 			0+0*8, 1+1*8, 2+0*8, 0+7*8,
 			2+2*8, 2+3*8, 2+4*8, 3+3*8,
 			3+4*8, 4+3*8, 4+4*8, 5+3*8,
